Fix unmarshaling CurrentWaterConsumption and CurrentEnergyConsumption.
Contrary to the documentation, these fields may be populated with
decimal numbers, not integers.

Documentation: https://www.miele.com/developer/swagger-ui/swagger.html#/Information/get_devices__deviceId__state
Observed value: "26.0"

+++ miele/device.go
@@ -119,12 +119,12 @@ type State struct {
 	PlateStep       []LocalizedValue `json:"plateStep"`
 	EcoFeedback     struct {
 		CurrentWaterConsumption struct {
-			Unit  string `json:"unit"`
-			Value int    `json:"value"`
+			Unit  string  `json:"unit"`
+			Value float64 `json:"value"`
 		} `json:"currentWaterConsumption"`
 		CurrentEnergyConsumption struct {
-			Unit  string `json:"unit"`
-			Value int    `json:"value"`
+			Unit  string  `json:"unit"`
+			Value float64 `json:"value"`
 		} `json:"currentEnergyConsumption"`
 		WaterForecast  float64 `json:"waterForecast"`
 		EnergyForecast float64 `json:"energyForecast"`
